Refactor generation orchestrator store into modules
