+++ app/frontend/src/features/generation/stores/orchestrator/watcherRegistry.ts
@@ -1,0 +1,23 @@
+import type { WatchStopHandle } from 'vue';
+
+export const createWatcherRegistry = () => {
+  let watchers: WatchStopHandle[] = [];
+
+  const register = (stopHandle: WatchStopHandle): void => {
+    watchers.push(stopHandle);
+  };
+
+  const stopAll = (): void => {
+    watchers.forEach((stop) => {
+      stop();
+    });
+    watchers = [];
+  };
+
+  return {
+    register,
+    stopAll,
+  };
+};
+
+export type WatcherRegistry = ReturnType<typeof createWatcherRegistry>;
